requirements-prod.txt
# Core FastAPI and server (Python 3.13 compatible)
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.11.0
pydantic-settings==2.6.1

# Google Cloud dependencies
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.18.1
firebase-admin==6.4.0
google-genai==0.3.2
google-cloud-aiplatform==1.38.1

# HTTP and networking
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Async support
anyio==4.2.0

# Logging and monitoring
google-cloud-logging==3.8.0

# JSON schema validation
jsonschema==4.20.0

# Date utilities
python-dateutil==2.8.2